Separate options.rb into two files.
#Amber-082

Options.rb is now seprated into two files, namely:
optins.rb and command_line_options.rb.  All rpecs
are update to reflect new files.
