+++ lib/amber/version.rb
@@ -11,5 +11,5 @@
 #     The next integer.  This number never resets.
 #
 module Amber
-  VERSION = '1.2.179'.freeze
+  VERSION = '1.2.180'.freeze
 end
