+++ spec/options/obliterate_spec.rb
@@ -1,5 +1,6 @@
 require 'rspec'
 require_relative '../../lib/amber/cli/options'
+require_relative '../../lib/amber/cli/command_line_options'
 # ------------------------------------------------------------------------------
 # obliterate options
 # [-O | --obliterate]
